+++ common/database/db_test.go
@@ -1,19 +1,19 @@
 package database_test
 
 import (
+	"scroll-tech/common/testcontainers/l1"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 
 	"scroll-tech/common/database"
-	"scroll-tech/common/testcontainers"
 	"scroll-tech/common/version"
 )
 
 func TestDB(t *testing.T) {
 	version.Version = "v4.1.98-aaa-bbb-ccc"
 
-	testApps := testcontainers.NewTestcontainerApps()
+	testApps := l1.NewTestcontainerApps()
 	assert.NoError(t, testApps.StartPostgresContainer())
 
 	db, err := testApps.GetGormDBClient()
