+++ common/testcontainers/l1/testcontainers.go
@@ -1,4 +1,4 @@
-package testcontainers
+package l1
 
 import (
 	"context"
