+++ common/testcontainers/l1/testcontainers_test.go
@@ -1,4 +1,4 @@
-package testcontainers
+package l1
 
 import (
 	"testing"
